requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1
django-recaptcha==4.0.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
django-taggit==6.1.0
filelock==3.15.4
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pytesseract==0.3.10
setuptools==70.1.1
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
wheel==0.43.0
